requirements.txt
# Python project dependencies for SmartEventAdder
# Google APIs (Calendar + Gmail)
google-api-python-client==2.181.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# Environment variables
python-dotenv==1.1.1

# AI Integration (for event parsing)
google-cloud-aiplatform==1.113.0

# Date/time handling
python-dateutil==2.9.0.post0

# HTTP requests (if needed for API calls)
requests==2.32.5

# JSON handling and data processing
jsonschema==4.25.1

# Testing dependencies
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0